Add an option to disable the new nicer ui menus for highlights and bookmarks
The new highlights and bookmark menus may be slower for extremely large
number of highlights and bookmarks. While the performance is definitely
enough for most users, some users might prefer the speed of the older
menus so we added a `fancy_ui_menus` which can be used to disable the
new fancy menus.